requirements.txt
fastapi
uvicorn[standard]
celery[redis]
pillow
aiofiles
python-multipart   
redis
amqp
pytesseract
